refs #33703 unit test
